Frontend code refactoring for better code organization
