+++ app/frontend/pages/__init__.py
@@ -1,0 +1,1 @@
+# Pages package
